+++ supertokens_python/recipe/emailpassword/api/email_exists.py
@@ -22,7 +22,7 @@ if TYPE_CHECKING:
     )
 
 from supertokens_python.exceptions import raise_bad_input_exception
-from supertokens_python.utils import send_200_response
+from supertokens_python.utils import send_200_response, default_user_context
 
 
 async def handle_email_exists_api(
@@ -33,6 +33,9 @@ async def handle_email_exists_api(
     email = api_options.request.get_query_param("email")
     if email is None:
         raise_bad_input_exception("Please provide the email as a GET param")
+    user_context = default_user_context(api_options.request)
 
-    response = await api_implementation.email_exists_get(email, api_options, {})
+    response = await api_implementation.email_exists_get(
+        email, api_options, user_context
+    )
     return send_200_response(response.to_json(), api_options.response)