+++ supertokens_python/recipe/emailpassword/api/generate_password_reset_token.py
@@ -22,7 +22,7 @@ if TYPE_CHECKING:
     )
 
 from supertokens_python.exceptions import raise_bad_input_exception
-from supertokens_python.utils import send_200_response
+from supertokens_python.utils import default_user_context, send_200_response
 
 from .utils import validate_form_fields_or_throw_error
 
@@ -40,8 +40,9 @@ async def handle_generate_password_reset_token_api(
         api_options.config.reset_password_using_token_feature.form_fields_for_generate_token_form,
         form_fields_raw,
     )
-    response = await api_implementation.generate_password_reset_token_post(
-        form_fields, api_options, {}
-    )
+    user_context = default_user_context(api_options.request)
 
+    response = await api_implementation.generate_password_reset_token_post(
+        form_fields, api_options, user_context
+    )
     return send_200_response(response.to_json(), api_options.response)