+++ supertokens_python/recipe/openid/api/open_id_discovery_configuration_get.py
@@ -12,7 +12,7 @@
 # License for the specific language governing permissions and limitations
 # under the License.
 from supertokens_python.recipe.openid.interfaces import APIInterface, APIOptions
-from supertokens_python.utils import send_200_response
+from supertokens_python.utils import default_user_context, send_200_response
 
 from ..interfaces import OpenIdDiscoveryConfigurationGetResponse
 
@@ -23,8 +23,10 @@ async def open_id_discovery_configuration_get(
     if api_implementation.disable_open_id_discovery_configuration_get:
         return None
 
+    user_context = default_user_context(api_options.request)
+
     result = await api_implementation.open_id_discovery_configuration_get(
-        api_options, {}
+        api_options, user_context
     )
 
     if isinstance(result, OpenIdDiscoveryConfigurationGetResponse):