+++ supertokens_python/recipe/passwordless/api/email_exists.py
@@ -13,7 +13,7 @@
 # under the License.
 from supertokens_python.exceptions import raise_bad_input_exception
 from supertokens_python.recipe.passwordless.interfaces import APIInterface, APIOptions
-from supertokens_python.utils import send_200_response
+from supertokens_python.utils import default_user_context, send_200_response
 
 
 async def email_exists(api_implementation: APIInterface, api_options: APIOptions):
@@ -24,5 +24,7 @@ async def email_exists(api_implementation: APIInterface, api_options: APIOptions
     if email is None:
         raise_bad_input_exception("Please provide the email as a GET param")
 
-    result = await api_implementation.email_exists_get(email, api_options, {})
+    user_context = default_user_context(api_options.request)
+
+    result = await api_implementation.email_exists_get(email, api_options, user_context)
     return send_200_response(result.to_json(), api_options.response)