+++ supertokens_python/recipe/passwordless/api/phone_number_exists.py
@@ -13,7 +13,7 @@
 # under the License.
 from supertokens_python.exceptions import raise_bad_input_exception
 from supertokens_python.recipe.passwordless.interfaces import APIInterface, APIOptions
-from supertokens_python.utils import send_200_response
+from supertokens_python.utils import default_user_context, send_200_response
 
 
 async def phone_number_exists(
@@ -26,7 +26,9 @@ async def phone_number_exists(
     if phone_number is None:
         raise_bad_input_exception("Please provide the phoneNumber as a GET param")
 
+    user_context = default_user_context(api_options.request)
+
     result = await api_implementation.phone_number_exists_get(
-        phone_number, api_options, {}
+        phone_number, api_options, user_context
     )
     return send_200_response(result.to_json(), api_options.response)