+++ supertokens_python/recipe/passwordless/api/resend_code.py
@@ -13,7 +13,7 @@
 # under the License.
 from supertokens_python.exceptions import raise_bad_input_exception
 from supertokens_python.recipe.passwordless.interfaces import APIInterface, APIOptions
-from supertokens_python.utils import send_200_response
+from supertokens_python.utils import default_user_context, send_200_response
 
 
 async def resend_code(api_implementation: APIInterface, api_options: APIOptions):
@@ -32,8 +32,9 @@ async def resend_code(api_implementation: APIInterface, api_options: APIOptions)
 
     pre_auth_session_id = body["preAuthSessionId"]
     device_id = body["deviceId"]
+    user_context = default_user_context(api_options.request)
 
     result = await api_implementation.resend_code_post(
-        device_id, pre_auth_session_id, api_options, {}
+        device_id, pre_auth_session_id, api_options, user_context
     )
     return send_200_response(result.to_json(), api_options.response)