+++ supertokens_python/recipe/thirdparty/api/apple_redirect.py
@@ -18,6 +18,8 @@ from typing import TYPE_CHECKING
 if TYPE_CHECKING:
     from supertokens_python.recipe.thirdparty.interfaces import APIInterface, APIOptions
 
+from supertokens_python.utils import default_user_context
+
 
 async def handle_apple_redirect_api(
     api_implementation: APIInterface, api_options: APIOptions
@@ -28,8 +30,11 @@ async def handle_apple_redirect_api(
 
     code = body["code"] if "code" in body else ""
     state = body["state"] if "state" in body else ""
+    user_context = default_user_context(api_options.request)
 
     # this will redirect the user...
-    await api_implementation.apple_redirect_handler_post(code, state, api_options, {})
+    await api_implementation.apple_redirect_handler_post(
+        code, state, api_options, user_context
+    )
 
     return api_options.response