+++ supertokens_python/recipe/thirdparty/api/authorisation_url.py
@@ -22,7 +22,7 @@ if TYPE_CHECKING:
     from supertokens_python.recipe.thirdparty.provider import Provider
 
 from supertokens_python.exceptions import raise_bad_input_exception
-from supertokens_python.utils import send_200_response
+from supertokens_python.utils import default_user_context, send_200_response
 
 
 async def handle_authorisation_url_api(
@@ -44,6 +44,9 @@ async def handle_authorisation_url_api(
             + third_party_id
             + " seems to be missing from the backend configs."
         )
+    user_context = default_user_context(api_options.request)
 
-    result = await api_implementation.authorisation_url_get(provider, api_options, {})
+    result = await api_implementation.authorisation_url_get(
+        provider, api_options, user_context
+    )
     return send_200_response(result.to_json(), api_options.response)