+++ CHANGELOG.md
@@ -6,14 +6,57 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 - Update tests to cover `resend_code` feature in `passwordless` and `thirdpartypasswordless` recipe.
-
 - Update usermetadata tests to ensure that utf8 chars are supported.
-
-# Fixes
-- Fixed handling of unicode characters in usermetadata (and emails, roles, session/access token payload data)
-
 - Mark tests as skipped if core version requirements are not met.
 - Use [black](https://github.com/psf/black) instead of `autopep8` to format code.
+
+### Features:
+- Adds default `user_context` for API calls that contains the request object. It can be used in APIs / functions override like this:
+
+```python
+def apis_override_email_password(param: APIInterface):
+    og_sign_in_post = param.sign_in_post
+
+    async def sign_in_post(
+        form_fields: List[FormField],
+        api_options: APIOptions,
+        user_context: Dict[str, Any],
+    ):
+        req = user_context.get("_default", {}).get("request")
+        if req:
+            # do something with the request
+
+        return await og_sign_in_post(form_fields, api_options, user_context)
+
+    param.sign_in_post = sign_in_post
+    return param
+
+def functions_override_email_password(param: RecipeInterface):
+    og_sign_in = param.sign_in
+
+    async def sign_in(email: str, password: str, user_context: Dict[str, Any]):
+        req = user_context.get("_default", {}).get("request")
+        if req:
+            # do something with the request
+
+        return await og_sign_in(email, password, user_context)
+
+    param.sign_in = sign_in
+    return param
+
+init(
+    ...,
+    recipe_list=[
+        emailpassword.init(
+            override=emailpassword.InputOverrideConfig(
+                apis=apis_override_email_password,
+                functions=functions_override_email_password,
+            )
+        ),
+        session.init(),
+    ],
+)
+```
 
 ### Breaking change
 - Change request arg type in session recipe functions from Any to BaseRequest.